fixed layout page to only display layout. shifted components to page.tsx
